requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
scipy==1.11.4
jupyter==1.0.0
ipykernel==6.27.1
python-dotenv==1.0.0
pyyaml==6.0.1
kaleido==0.2.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
